Move Tx/UTxO serialization tests to hydra-tx
These are tests that were made to cover our orphan instances of
To/FromJSON and To/FromCBOR for cardano-api Tx and UTxO types.
