+++ hydra-tx/golden/ReasonablySized (Tx ConwayEra).json
